+++ kong/plugins/request-transformer/access.lua
@@ -33,7 +33,6 @@ local pl_copy_table = pl_tablex.deepcopy
 
 local _M = {}
 local template_cache = setmetatable( {}, { __mode = "k" })
-local template_environment
 
 local DEBUG = ngx.DEBUG
 local CONTENT_LENGTH = "content-length"
@@ -77,54 +76,7 @@ local function get_content_type(content_type)
   end
 end
 
--- meta table for the sandbox, exposing lazily loaded values
-local __meta_environment = {
-  __index = function(self, key)
-    local lazy_loaders = {
-      headers = function(self)
-        return get_headers() or EMPTY
-      end,
-      query_params = function(self)
-        return get_uri_args() or EMPTY
-      end,
-      uri_captures = function(self)
-        return (ngx.ctx.router_matches or EMPTY).uri_captures or EMPTY
-      end,
-      shared = function(self)
-        return ((kong or EMPTY).ctx or EMPTY).shared or EMPTY
-      end,
-    }
-    local loader = lazy_loaders[key]
-    if not loader then
-      -- we don't have a loader, so just return nothing
-      return
-    end
-    -- set the result on the table to not load again
-    local value = loader()
-    rawset(self, key, value)
-    return value
-  end,
-  __newindex = function(self)
-    error("This environment is read-only.")
-  end,
-}
-
-template_environment = setmetatable({
-  -- here we can optionally add functions to expose to the sandbox, eg:
-  -- tostring = tostring,  -- for example
-  -- because headers may contain array elements such as duplicated headers
-  -- type is a useful function in these cases. See issue #25.
-  type = type,
-}, __meta_environment)
-
-local function clear_environment(conf)
-  rawset(template_environment, "headers", nil)
-  rawset(template_environment, "query_params", nil)
-  rawset(template_environment, "uri_captures", nil)
-  rawset(template_environment, "shared", nil)
-end
-
-local function param_value(source_template, config_array)
+local function param_value(source_template, config_array, template_env)
   if not source_template or source_template == "" then
     return nil
   end
@@ -146,10 +98,10 @@ local function param_value(source_template, config_array)
     compiled_templates[source_template] = compiled_template
   end
 
-  return compiled_template:render(template_environment)
+  return compiled_template:render(template_env)
 end
 
-local function iter(config_array)
+local function iter(config_array, template_env)
   return function(config_array, i, previous_name, previous_value)
     i = i + 1
     local current_pair = config_array[i]
@@ -163,7 +115,7 @@ local function iter(config_array)
       return i, current_name
     end
 
-    local res, err = param_value(current_value, config_array)
+    local res, err = param_value(current_value, config_array, template_env)
     if err then
       return error("[request-transformer] failed to render the template " ..
                    current_value .. ", error:" .. err)
@@ -189,14 +141,14 @@ local function append_value(current_value, value)
   end
 end
 
-local function transform_headers(conf)
+local function transform_headers(conf, template_env)
   local headers = get_headers()
   local headers_to_remove = {}
 
   headers.host = nil
 
   -- Remove header(s)
-  for _, name, value in iter(conf.remove.headers) do
+  for _, name, value in iter(conf.remove.headers, template_env) do
     name = name:lower()
     if headers[name] then
       headers[name] = nil
@@ -205,7 +157,7 @@ local function transform_headers(conf)
   end
 
   -- Rename headers(s)
-  for _, old_name, new_name in iter(conf.rename.headers) do
+  for _, old_name, new_name in iter(conf.rename.headers, template_env) do
     old_name = old_name:lower()
     local value = headers[old_name]
     if value then
@@ -216,7 +168,7 @@ local function transform_headers(conf)
   end
 
   -- Replace header(s)
-  for _, name, value in iter(conf.replace.headers) do
+  for _, name, value in iter(conf.replace.headers, template_env) do
     name = name:lower()
     if headers[name] or name == HOST then
       headers[name] = value
@@ -224,14 +176,14 @@ local function transform_headers(conf)
   end
 
   -- Add header(s)
-  for _, name, value in iter(conf.add.headers) do
+  for _, name, value in iter(conf.add.headers, template_env) do
     if not headers[name] and name:lower() ~= HOST then
       headers[name] = value
     end
   end
 
   -- Append header(s)
-  for _, name, value in iter(conf.append.headers) do
+  for _, name, value in iter(conf.append.headers, template_env) do
     local name_lc = name:lower()
 
     if name_lc ~= HOST and name ~= name_lc and headers[name] ~= nil then
@@ -254,7 +206,7 @@ local function transform_headers(conf)
   set_headers(headers)
 end
 
-local function transform_querystrings(conf)
+local function transform_querystrings(conf, template_env)
 
   if not (#conf.remove.querystring > 0 or #conf.rename.querystring > 0 or
           #conf.replace.querystring > 0 or #conf.add.querystring > 0 or
@@ -262,41 +214,41 @@ local function transform_querystrings(conf)
     return
   end
 
-  local querystring = pl_copy_table(template_environment.query_params)
+  local querystring = pl_copy_table(template_env.query_params)
 
   -- Remove querystring(s)
-  for _, name, value in iter(conf.remove.querystring) do
+  for _, name, value in iter(conf.remove.querystring, template_env) do
     querystring[name] = nil
   end
 
   -- Rename querystring(s)
-  for _, old_name, new_name in iter(conf.rename.querystring) do
+  for _, old_name, new_name in iter(conf.rename.querystring, template_env) do
     local value = querystring[old_name]
     querystring[new_name] = value
     querystring[old_name] = nil
   end
 
-  for _, name, value in iter(conf.replace.querystring) do
+  for _, name, value in iter(conf.replace.querystring, template_env) do
     if querystring[name] then
       querystring[name] = value
     end
   end
 
   -- Add querystring(s)
-  for _, name, value in iter(conf.add.querystring) do
+  for _, name, value in iter(conf.add.querystring, template_env) do
     if not querystring[name] then
       querystring[name] = value
     end
   end
 
   -- Append querystring(s)
-  for _, name, value in iter(conf.append.querystring) do
+  for _, name, value in iter(conf.append.querystring, template_env) do
     querystring[name] = append_value(querystring[name], value)
   end
   set_uri_args(querystring)
 end
 
-local function transform_json_body(conf, body, content_length)
+local function transform_json_body(conf, body, content_length, template_env)
   local removed, renamed, replaced, added, appended = false, false, false, false, false
   local content_length = (body and #body) or 0
   local parameters = parse_json(body)
@@ -308,14 +260,14 @@ local function transform_json_body(conf, body, content_length)
   end
 
   if content_length > 0 and #conf.remove.body > 0 then
-    for _, name, value in iter(conf.remove.body) do
+    for _, name, value in iter(conf.remove.body, template_env) do
       parameters[name] = nil
       removed = true
     end
   end
 
   if content_length > 0 and #conf.rename.body > 0 then
-    for _, old_name, new_name in iter(conf.rename.body) do
+    for _, old_name, new_name in iter(conf.rename.body, template_env) do
       local value = parameters[old_name]
       parameters[new_name] = value
       parameters[old_name] = nil
@@ -324,7 +276,7 @@ local function transform_json_body(conf, body, content_length)
   end
 
   if content_length > 0 and #conf.replace.body > 0 then
-    for _, name, value in iter(conf.replace.body) do
+    for _, name, value in iter(conf.replace.body, template_env) do
       if parameters[name] then
         parameters[name] = value
         replaced = true
@@ -333,7 +285,7 @@ local function transform_json_body(conf, body, content_length)
   end
 
   if #conf.add.body > 0 then
-    for _, name, value in iter(conf.add.body) do
+    for _, name, value in iter(conf.add.body, template_env) do
       if not parameters[name] then
         parameters[name] = value
         added = true
@@ -342,7 +294,7 @@ local function transform_json_body(conf, body, content_length)
   end
 
   if #conf.append.body > 0 then
-    for _, name, value in iter(conf.append.body) do
+    for _, name, value in iter(conf.append.body, template_env) do
       local old_value = parameters[name]
       parameters[name] = append_value(old_value, value)
       appended = true
@@ -354,19 +306,19 @@ local function transform_json_body(conf, body, content_length)
   end
 end
 
-local function transform_url_encoded_body(conf, body, content_length)
+local function transform_url_encoded_body(conf, body, content_length, template_env)
   local renamed, removed, replaced, added, appended = false, false, false, false, false
   local parameters = decode_args(body)
 
   if content_length > 0 and #conf.remove.body > 0 then
-    for _, name, value in iter(conf.remove.body) do
+    for _, name, value in iter(conf.remove.body, template_env) do
       parameters[name] = nil
       removed = true
     end
   end
 
   if content_length > 0 and #conf.rename.body > 0 then
-    for _, old_name, new_name in iter(conf.rename.body) do
+    for _, old_name, new_name in iter(conf.rename.body, template_env) do
       local value = parameters[old_name]
       parameters[new_name] = value
       parameters[old_name] = nil
@@ -375,7 +327,7 @@ local function transform_url_encoded_body(conf, body, content_length)
   end
 
   if content_length > 0 and #conf.replace.body > 0 then
-    for _, name, value in iter(conf.replace.body) do
+    for _, name, value in iter(conf.replace.body, template_env) do
       if parameters[name] then
         parameters[name] = value
         replaced = true
@@ -384,7 +336,7 @@ local function transform_url_encoded_body(conf, body, content_length)
   end
 
   if #conf.add.body > 0 then
-    for _, name, value in iter(conf.add.body) do
+    for _, name, value in iter(conf.add.body, template_env) do
       if parameters[name] == nil then
         parameters[name] = value
         added = true
@@ -393,7 +345,7 @@ local function transform_url_encoded_body(conf, body, content_length)
   end
 
   if #conf.append.body > 0 then
-    for _, name, value in iter(conf.append.body) do
+    for _, name, value in iter(conf.append.body, template_env) do
       local old_value = parameters[name]
       parameters[name] = append_value(old_value, value)
       appended = true
@@ -405,12 +357,12 @@ local function transform_url_encoded_body(conf, body, content_length)
   end
 end
 
-local function transform_multipart_body(conf, body, content_length, content_type_value)
+local function transform_multipart_body(conf, body, content_length, content_type_value, template_env)
   local removed, renamed, replaced, added, appended = false, false, false, false, false
   local parameters = multipart(body and body or "", content_type_value)
 
   if content_length > 0 and #conf.rename.body > 0 then
-    for _, old_name, new_name in iter(conf.rename.body) do
+    for _, old_name, new_name in iter(conf.rename.body, template_env) do
       if parameters:get(old_name) then
         local value = parameters:get(old_name).value
         parameters:set_simple(new_name, value)
@@ -421,14 +373,14 @@ local function transform_multipart_body(conf, body, content_length, content_type
   end
 
   if content_length > 0 and #conf.remove.body > 0 then
-    for _, name, value in iter(conf.remove.body) do
+    for _, name, value in iter(conf.remove.body, template_env) do
       parameters:delete(name)
       removed = true
     end
   end
 
   if content_length > 0 and #conf.replace.body > 0 then
-    for _, name, value in iter(conf.replace.body) do
+    for _, name, value in iter(conf.replace.body, template_env) do
       if parameters:get(name) then
         parameters:delete(name)
         parameters:set_simple(name, value)
@@ -438,7 +390,7 @@ local function transform_multipart_body(conf, body, content_length, content_type
   end
 
   if #conf.add.body > 0 then
-    for _, name, value in iter(conf.add.body) do
+    for _, name, value in iter(conf.add.body, template_env) do
       if not parameters:get(name) then
         parameters:set_simple(name, value)
         added = true
@@ -451,7 +403,7 @@ local function transform_multipart_body(conf, body, content_length, content_type
   end
 end
 
-local function transform_body(conf)
+local function transform_body(conf, template_env)
   local content_type_value = get_header(CONTENT_TYPE)
   local content_type = get_content_type(content_type_value)
   if content_type == nil or #conf.rename.body < 1 and
@@ -461,16 +413,19 @@ local function transform_body(conf)
   end
 
   -- Call req_read_body to read the request body first
-  local body = get_raw_body()
+  local body, err = get_raw_body()
+  if err then
+    kong.log.warn(err)
+  end
   local is_body_transformed = false
   local content_length = (body and #body) or 0
 
   if content_type == ENCODED then
-    is_body_transformed, body = transform_url_encoded_body(conf, body, content_length)
+    is_body_transformed, body = transform_url_encoded_body(conf, body, content_length, template_env)
   elseif content_type == MULTI then
-    is_body_transformed, body = transform_multipart_body(conf, body, content_length, content_type_value)
+    is_body_transformed, body = transform_multipart_body(conf, body, content_length, content_type_value, template_env)
   elseif content_type == JSON then
-    is_body_transformed, body = transform_json_body(conf, body, content_length)
+    is_body_transformed, body = transform_json_body(conf, body, content_length, template_env)
   end
 
   if is_body_transformed then
@@ -512,10 +467,10 @@ local function transform_method(conf)
   end
 end
 
-local function transform_uri(conf)
+local function transform_uri(conf, template_env)
   if conf.replace.uri then
 
-    local res, err = param_value(conf.replace.uri, conf.replace)
+    local res, err = param_value(conf.replace.uri, conf.replace, template_env)
     if err then
       error("[request-transformer] failed to render the template " ..
         tostring(conf.replace.uri) .. ", error:" .. err)
@@ -531,12 +486,51 @@ local function transform_uri(conf)
 end
 
 function _M.execute(conf)
-  clear_environment()
-  transform_uri(conf)
+  -- meta table for the sandbox, exposing lazily loaded values
+  local __meta_environment = {
+    __index = function(self, key)
+      local lazy_loaders = {
+        headers = function(self)
+          return get_headers() or EMPTY
+        end,
+        query_params = function(self)
+          return get_uri_args() or EMPTY
+        end,
+        uri_captures = function(self)
+          return (ngx.ctx.router_matches or EMPTY).uri_captures or EMPTY
+        end,
+        shared = function(self)
+          return ((kong or EMPTY).ctx or EMPTY).shared or EMPTY
+        end,
+      }
+      local loader = lazy_loaders[key]
+      if not loader then
+        -- we don't have a loader, so just return nothing
+        return
+      end
+      -- set the result on the table to not load again
+      local value = loader()
+      rawset(self, key, value)
+      return value
+    end,
+    __newindex = function(self)
+      error("This environment is read-only.")
+    end,
+  }
+
+  local template_env = setmetatable({
+    -- here we can optionally add functions to expose to the sandbox, eg:
+    -- tostring = tostring,  -- for example
+    -- because headers may contain array elements such as duplicated headers
+    -- type is a useful function in these cases. See issue #25.
+    type = type,
+  }, __meta_environment)
+
+  transform_uri(conf, template_env)
   transform_method(conf)
-  transform_headers(conf)
-  transform_body(conf)
-  transform_querystrings(conf)
+  transform_headers(conf, template_env)
+  transform_body(conf, template_env)
+  transform_querystrings(conf, template_env)
 end
 
 return _M