+++ CHANGELOG.md
@@ -108,6 +108,11 @@
   Docker containers if it was not cleanly stopped. 
   [#10468](https://github.com/Kong/kong/pull/10468)
 
+#### Plugins
+- **Request-Transformer**: fix an issue where requests would intermittently
+  be proxied with incorrect query parameters.
+  [10539](https://github.com/Kong/kong/pull/10539)
+
 ### Changed
 
 #### Core